requirements.txt
requests-exoscale-auth>=1.1.2
boto3>=1.40.16
importlib>=1.0.4
pydantic>=2.0
python-dotenv>=1.0
requests>=2.32.3